This addition to jwst_lib.models provides the initial (throwaway) implementation of new data models for supporting the resample step (drizzling) for processing imaging data.  In particular, the AsnModel with corresponding asn schema and the DrizParsModel and schema for defining the FITS versions of ASN tables and DrizPars reference table (similar to MDRIZTAB for HST).
This revision is a pre-requisite for the changes being implemented for the resample pipeline step.

The implementation of the ASN schema and data model will be almost immediately replaced with a simpler JSON-based definition and schema. A new schema for validating a JSON formatted drizpars file may also be implemented shortly once we have experience with how to use that format with the input asn table.


git-svn-id: https://aeon.stsci.edu/ssb/svn/jwst/trunk/jwst_lib/models@2814 560b4ebf-6bc0-4cc5-b8e0-b136f69d22d4