Added code to read DQ_DEF extension from reference files and apply the dynamic_dq flag assignments.
DQ_DEF extension needs to be a table with 4 columns with the bit assignments
used in the DQ extension of the reference file:

BIT          Int32    Bit plane (0-32)
VALUE        Int32    2**BIT
NAME         S40      From dqflags.py
DESCRIPTION  S80      Not actually used in the code, but could be useful

So a row that looks like

BIT = 2
VALUE = 4
NAME = 'TELEGRAPH'
DESCRIPTION = 'Telegraph pixel'

has bitplane 2 (i.e. value 4) used to flag TELEGRAPH pixels.  When the
reference file is turned into a model, the dq attribute of the model will
have those pixels that AND with 4 in the original data OR'd with the
value of TELEGRAPH in dqflags.py.



git-svn-id: https://aeon.stsci.edu/ssb/svn/jwst/trunk/jwst_lib/models@2924 560b4ebf-6bc0-4cc5-b8e0-b136f69d22d4